admission: move GrantCoordinator into its own file
Also bring along with it some closely related structs:
StoreGrantCoordinators, Options, GrantCoordinatorMetrics. The latter was
previously named GranterMetrics, but this feels more appropriate. Pure
code movement change.

Release note: None
